TNG-90: First working attempt at support for multiple "themes"
The design goal for this part of development was to support an arbitrarily
structured tree of tree of front-end themes. This is temporarily done with a
lot of simplifying assumptions, mainly that the themes are always located at
directories on the local disk. The longer-term goal is to support a very
general case of where the themes come from, such as from a service call. The
classes ThemeAccessor and JournalTemplateLoader are tightly coupled to the
assumption that themes always come from the local disk, and will probably
become obsolete; the classes ThemeTree and DelegatingTemplateLoader are more
general and hopefully will remain useful.

It is still an open question whether we want to keep "theme inheritance"
granular on the level of files.