Implement basic building of accordion sections in JS, on page load
In order for the accordion sections to work, the page's HTML will have to
be structured differently than is currently done by the XSLT transform.
This is an experiment in the direction of changing the structure in
JavaScript, after rather than during the transform. The two main advantages
are: (1) ease of implementation compared to editing XSLT and (2) graceful
degradation in case the user browses without JavaScript. (Explanation of
(2): this way, the page can be rendered with all text in place, and it is
hidden behind JavaScript expanders only if JavaScript is running in the
first place.)

This is an early pass. There are lots of blemishes in the text as rendered.
The '.abstract' and '.articleinfo' sections are omitted entirely.