Change terminology to use "site" instead of "journal" on the Wombat side
Wombat so far has overloaded the word "journal" too much. In the system at
large, it currently has two meanings:

    (A) A publication that contains a set of articles. This is the
        definition used in Ambra and Rhino. These journal objects are
        accessed through the API's "journals" namespace.

    (B) A set of front-end pages associated with a theme. This is the
        definition Wombat has been using so far.

The overloading is confusing in itself, and will become more so when
journals (B) have a many-to-one relationship with journals (A). For
example, "MobilePlosOne" and "DesktopPlosOne" might be distinct journals
(B) one day, but they would both correspond to the single journal (A) that
is PLOS ONE.

Therefore, Meaning B of "journal" is deprecated and replaced with the word
"site". This is consistent already with how we have been speaking of the
"mobile site" and "desktop site". "Journal" now should have Meaning A
across all components.

The awkward term "publication key" was introduced in Wombat a few commits
ago, to refer to journal (A) keys in a context where "journal" normally
meant B. It is now replaced with "journal key" in the sense of A.

This commit contains a lot of renaming, but should cause no functional
changes.