AMEC-1898: In JavaScript, kludge out figure descriptions' links to nowhere
Links that are supposed to go to another section of the article sometimes
appear in figure descriptions. With the current transform, these show up on
the figure page as links to nonexistent anchors on the same page.

As a more expedient fix than changing the transform, have the JavaScript on
page load look for such links inside figure description text and remove
them. A less hackish fix would be nice.