AMEC-2193: Prevent 500 errors when empty strings are given as URL params
Spring's RequestParam mapping checks only that required parameters are
present (non-null) but allows empty strings, so these are checked in the
controller method bodies.

Introduced NotFoundException as a general superclass of
ArticleNotFoundException. The distinction may be useful in the future for
displaying different types of 404 page.

Tweaked functionality in DeserializedJsonUtil.readField, which isn't used
elsewhere.