AMEC-2275: Display amendments as links to separate articles
This both introduces new support for retractions and EOCs and replaces the
support for old-style corrections (annotation-based) with new-style
corrections (separate XML articles).

All three amendment types now appear on the front end as links to the full
amendment article. These links indicate the amendment type but (unlike
old-style corrections) otherwise do not contain any text from it. This is
less featureful than the support for annotation-based corrections, which
did show the correction body in the corrected article.