AMEC-2311: Add front-end support for homepage sections
Previously, the controller supported homepage sections ("recent",
"popular", and "in the news") but there was no front-end code to support
them (it was only in the PLOS themes). Add section-switching buttons to the
homepage code in the webapp, but show them only if a theme config enables
more than one section.