AMEC-2335: Allow caching requests on unsupported services
If a request is made to cache results from a service that does not support
the If-Modified-Since header, silently carry through the request and skip
inserting the result into the cache (since we wouldn't be able to retrieve
it anyway).

This is best for an experimental phase in which we want to to aggressively
cache lots of things without worrying about matching client requests with
the services that support the If-Modified-Since header. Otherwise, it would
be better to error out (or at least log warnings) if we expect cacheability
where it isn't supported.