AMEC-2461: Refactor controller method for unique asset files
At the time I wrote this code, there was a performance benefit to having
two different service queries: one for when we had an actual file ID, and
one for when we had only an asset ID with a unique file. However, we now
query for asset metadata anyway (in order to validate article visibility),
meaning we have the file ID in all cases. So it is much simpler to query
the service API using only the file ID.

Refactored to do it this simpler way, allowing the removal of the "?unique"
query from the service API side.