DPRO-21: Refactor SiteRequestScheme to pull in site-building logic
The buildLink part of the former interface does not compose as cleanly as
the isForSite method. We want to combine several interfaces that describe
the site definition -- path token plus headers -- though only the former
affects the way links are built.

As of this commit, assume that the class is concerned only with building
absolute links to pages within the same site. Therefore, only the "path
token" part of definition needs special handling, which is done in the
buildLink method of the SiteRequestScheme path definition. The "is for
site" logic remains as an interface (SiteRequestPredicate) that can be
composed with simple "and" logic.