DPRO-180: Have desktop theme's search.ftl tolerate null legacy.urlPrefix
It now points at the search controller's legacy mode (and links to the
legacy server's advanced search) only if legacy.urlPrefix is defined.
Otherwise, it shows a form for the webapp's own simple search (as in the
mobile theme).

This required changing VariableLookupDirective's contract slightly to
permit null values when they are assigned to a FreeMarker loopvar. Nulls
still cause an error if you attempt to print them directly to the page.