DPRO-290: Make relative links to assets absolute
Fixes a bug where the desktop homepage is viewed without a trailing slash
in the browser's address (ordinarily impossible because we 302 from there,
but may come up depending on what the web server is doing in front of us).
It would also have been a bug when the code was reused in non-homepage
contexts.