DPRO-455: Don't fetch or display amendment body on corrections
Previously the controller would query for the body whether it would be
shown or not. This change moves the definition of whether to display the
body per amendment type from the front end to the controller. It would be
better to allow it to be configured per theme (and still prevent the
controller from making a wasted query).