DPRO-455: Prefer ArticleTransformService over XmlExcerptTransformation
This seems to produce working results, despite differing from the legacy
implementation. The special cases not covered in the transform XML were
likely specific to author info, not amendment notices.

Changed contract on ArticleTransformService.transformExcerpt to permit a
null value on enclosingTag, for new use case.

Deleted XmlExcerptTransformation, which is now unused.