DPRO-435: Debug addition of class="toc-section" in transform
The element
    <xsl:attribute name="class">toc-section</xsl:attribute>
was meant in several places to add a "toc-section" class to the enclosing
div, but it actually overwrites the div's existing class. This caused
class="abstract", class="section", and a few other things to disappear.

Instead, declare the "toc-section" class on the div itself, or combine with
another <xsl:attribute name="class"> element if one exists on the same
template (if there are two, they clobber each other).