DPRO-448: Fix file size calc
In the spirit of Halloween, the dev team is proud to present The Story that
Wouldn't Die! Like a circling vulture, it spirals ever closer and closer to
the behavior of the reference implementation. Will it ever get there?

Rounds file sizes to nearest integer if shown in units smaller than MB.