DPRO-574- per code review a more "fun" constant
I made this 30 and put it in a variable so it was clear. I arrived at
this by taking the width and subtracting the padding. this is 3 pixels
smaller than the original number but seems safer to me. It allows for
super wide letters however, removing the wiggle room does mean that
more of the letters may be cut off.