DPRO-706 - fixing scrolling of references for iPad
It appears that making the body fixed while having the fi viewer open
allows things to scroll automatically on iPad. Alternately this also
fixed the problem of the fig-viewer appearing slightly smaller than the
viewport on ipads.
I am assuming that since this is how lightbox works on Ambra, this is
how we want it to work.