DPRO-788: Don't show a 404 page if only article XML is missing
Removed a "catch (EntityNotFoundException)" block. At that point, article
existence has already been checked, so an EntityNotFoundException indicates
an invalid data condition where the article is in Rhino's database but the
XML file is missing. In this case it is more useful to indicate a
server-side error than to report a 404.