DPRO-1289: render journal-specific links for search results.
This was made a little complicated since there is at least one case
where the first journalKey in the cross_published_journal_key field
returned by solr is PLoSCollections.  We always want to return the
non-collections URL.  I piggybacked on some code that previously
existed for determining whether a given site corresponds to a
collection.

I also took the opportunity to build out a more realistic
TestThemeTree, which I've wanted to do for a while.