DPRO-1454: Throw exception where we can't build a correct relative link
There is an irreparable bug in the prior implementation, where we can't
link to a site on the "default" hostname if the origin site is mapped to a
different hostname, because it is impossible to recover the "default"
hostname from the request object. This commit introduces a requirement that
either every site has a configured hostname, or no site does.