DPRO-1508: Apply port from X-Forwarded-Host when building links
The problem is that the X-Forwarded-Host header may provide a port number
that we must use when eventually building a link; if it does not, we must
leave the port number off the link and *not* use the request's port.

This change encapsulates the hostname-plus-optional-port in the new
ClientEndpoint class. It absorbs a little display logic from the Link class
about not showing the default port.

As a bonus, this fixes a bug in HostPredicate where a header value with a
port might not get correctly matched against a configured hostname.