DPRO-1502: Remove global site look-ups by journal key
Looking up a site by journal key isn't generally safe, because it returns a
set of sites with no way to distinguish between them. (Compare to
Theme.resolveForeignJournalKey.) The one place it was used (outside of unit
tests) was to look up the journal's display name.

Replaced this with a narrower way to look up the unique journal name from a
journal key. Journal names are now required values, and we explicitly check
that they are consistent among all sites with the same journal key.