DPRO-1613: Resolve correctly if condition is for siteless handler
Fixes a bug where a request would not resolve to a siteless handler if
there happened to be a site predicate so broad that it matched the request
anyway.

Refactored out the map of PatternsRequestConditions, because it was
misleading in two out of the three cases we are handling. Replaced it with
three factory methods (forSiteless, forSingleSite, forSiteMap) that
separate out those cases more cleanly.