DPRO-1756 Move the browse subject area endpoint to SearchController
Unfortunately all of the logic for parsing search params and adding them to the model
are in the SearchController, so I moved the end point to be able to utilize them. Although
for browse we use "search" logic and maybe it really belongs in the SearchController,
I am not sure about using "browse" namespace.