DPRO-1910: Debug error-handling on taxonomy endpoint
This change prevents an error where "IllegalStateException:
getOutputStream() has already been called for this response" was logged to
the console, obscuring another error.

The fix is to request the service input stream before opening the output
stream. If an exception occurs on the service request, our normal
exception-handling code will attempt to open the output stream again.
So, we don't open the output stream until we're sure we can write to it.