DPRO-1899: Rearrange code into a more semantic order
Sorry for the banner comments. This is an incremental improvement in
readability I think. It would be great to break this class up into more
files but we would have to handle the wired dependencies somehow. It will
need to be refactored when we put things other than articles in feeds.