DPRO-2038: Default multi-value parameters to storage as lists
Search parameters with more than one value will always be persisted as a
list. This prevents the possibility of discarding the second value.

Assuming the input parameters are always valid, this will never save
invalid data. If the input parameters are weird, err on the side of saving
incorrectly formatted data over incomplete data.