DPRO-2389: Simplify ScholarlyWorkController
Removed TypeRedirect interface. It was intended to handle more complex
link-building cases, but as long as the input is constrained to be only a
ScholarlyWorkId, nothing more complex is really possible. It could
conceivably be needed if the "id" and "rev" params ever vary between
handlers, but we wouldn't want to do that anyway.