DPRO-2730: Clean up abstraction in ArticleResolutionService
The service now returns an ArticlePointer and is no longer concerned with
building an ApiAddress.

This removes an optimization where we can skip the remote call to get the
article overview, in cases where the RequestedDoiVersion has an ingestion
number and only an ApiAddress (not the canonical DOI) is needed. However,
the complexity cost would be substantial and it is likely to improve
performance only in low-traffic cases (assuming the ingestion number is
used mainly for QC purposes).