DPRO-2596: Avoid supplying a bogus 'id' parameter to volume handler
The "browse volumes" page is currently unusual in that it only supports
viewing all volumes at once, and doesn't take an 'id' parameter like the
other handlers. We could supply the parameter knowing that it will be
ignored, but that is misleading at best. We also might need to build links
in other ways than with a parameter named 'id' in the future. Replaced the
one-size-fits-all code for building links with more flexible lambdas.