DPRO-3073: Replace Site.isJournalSpecific with instanceof
In general, it would be preferable to avoid using instanceof. But there are
only two applicable places in non-test code and one of them requires
casting to JournalSite anyway.

If too many instanceof operators pile up, consider refactoring so that Site
has a method that returns an Optional object encapsulating the journalKey
and journalName.