DPRO-3156: Debug NPE on citation download when article has no description
Considered instead having XmlUtil.extractText return null for a null
argument, but that would be a little uglier and also inconsistent with
extractElement, etc.

Also considered adding similar checks in CitationDownloadServiceImpl for
title, since it is handled similarly to description. But title is expected
always to be present, so failing loudly is fine.