DPRO-3117: Add multi-site case
Exercises the way that CommentCensorServiceImplTest aggregates words that
are common to more than one site.

Using a DataProvider in this way might be overkill. It is a proof of
concept in case we need, say, different SiteSet values for individual test
cases. We may simplify it with a single SiteSet later as part of this task.