DPRO-3015: Replace defensive regex instead of 'https://dx.doi.org/'
It would not be correct just to replace this one with 'https://doi.org/'
because it is targeting data from an external service. This guards against
a style change on their end and makes it clear to our own maintainers that
it isn't just a straggler 'https://dx.doi.org/' that needs updating.