ENG-119 Fix problem with journal, article type faceting
Article type and journal are facets that should be ORed together, not
ANDed together. That is, when a user is attempting to facet, they want
articles in ONE OR Medicine, not both in ONE AND Medicine (which is
impossible). In order to support this, we need to exclude certain
facet queries from the faceting mechanism. That is, we want to return
facet counts as though the facet query did not happen, because we want
to return the counts that a user would get if they added an OR clause
to the facet query. Fortunately, solr supports this out of the box
using tagging/excluding filters.

For more details see:
https://lucene.apache.org/solr/guide/7_4/faceting.html#Faceting-TaggingandExcludingFilters